requirements.txt
requests
edge-tts
google-genai
selenium
beautifulsoup4
markdownify
boto3
psycopg2-binary